feat: login admin and user CRUD

+++ db/migrations/20231012133408-create-admin.js
@@ -16,7 +16,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       nomor_hp: {
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT(15)
       },
       createdAt: {
         allowNull: false,
